Last update for a while, I'm taking a hiatus. -Jaden
I'm possibly going to be moving out, so please bare with me here, okay? It's going to take a few weeks, possibly more to get internet again. Duke will still be here to answer questions and work on the Inventory system.

I made some changes to the interface, so that it looks more like Undertale hopefully hehe

Also added fixedsys fonts to use for the body, wherever Frisk is describing something in their POV.

Changed Gaster's pic so that he is actually grimacing now, lol!