Lots of bug fixes and updates! I'm back!
Still living in the same place, but we had an after-funeral get-together that we had to get the place fixed up for, and it took a full couple weeks or so.

Anyway!

I fixed up of the css code, tried to get the saves working better (there is now only one save slot, but you'll need to save to a file and delete your old save if you saved one before.

Added a ETHERTALE title image!

And many other various tweaks and improvements. This version is v0.0.2ALPHA.