Debating whether or not to add music to this game... I dunno how I'd get any music for it tbh
I did add a nice free-to-use sad piano solo, but I reallly need some sci-fi-esque electronic or dubstep ambient music (in the style of CORE probably, but ya know, slower but still memorable) to fill out the soundtrack.

Anyway.

This update I haven't added functionality for speech boxes for the E-T characters yet, but I have added the code to incorporate it in later builds. I've also made some art for Sans to put in the boxes' image box, which is to the left of the speech, as normal for an UT-AU game. ;)

Some of them have black backgrounds because the serious ones I didn't want to have a white outline around, heheh

Typed-js is a module for sugarcube, we'll see if it works for making the character's speaking speeds different, idk if it will... But I'm hopeful. I haven't looked at the documentation yet tho honestly lol

I think that's it. This release isn't really very playable tho, so play the release that I did before. It should be better for playability (hopefully). The first one is definitely playable, that's all I know XD

Besides that, I hope this stems the (admittedly not-very-existent) anticipation-!