Updating images and some inv code
Been trying to get my inventory code working enough to let Frisk equip a knife so I can do battles. Also got started implementing the character dialogue boxes. Sans' are done, will be doing Gaster's and Undyne's in the next few updates. It might take me a bit to get Undyne's art done, so expect just Gaster's in the next update after this one. ;)

I also made up a Tumblr dev blog for Ethertale, it's pretty empty rn but it'll be filled up with all of my images later to use as a reference for the online version of the game.
https://ethertale-fangame.tumblr.com

I think the downloadable version should still reference the images in the folder, though. :)

Also added some placeholder music. Alien in the Ether is going to be for the fights with aliens on the first planet, and creepily puzzled is obviously going to play while Frisk and the player try to figure out puzzles on the ship and various planets.

Welp, thanks for reading. :D