Lots of Art
Also going to be using Meters.min.js by Chapel instead of the healthbar.js by HiEv. The former is more square in shape, which fits the game more tbh.
Been working on talk-macro character expression heads, using Krita to do the art. I'll be working on getting the bugs out of the next push so that I can release the game again, a lot of stuff has changed and I'm hoping I can get the bugs out in a reasonable timeframe... X3