Made the inv system a bit easier for myself, as I can't Math(tm)
It's easier for me to not count every time I need to figure out which id is which lol

Also, added the fight buttons to be used later

Edited the README.md a bit, to make the instructions a bit easier to follow.

And that's it! For now at least.