Changed the tags in the places I want there to be a list of the available items
I also fixed getting the previous passage tags and so now whenever the player goes to a menu, the time won't continue after they return to the main game.

Thanks to Cyrus for the help with that!

I've also edited the game to make it so that it's aligning with the map that I made a while ago for the SS Bonetrousle.